requirements.txt
# Core Python libraries
pathlib2>=2.3.7; python_version < "3.4"

# Data handling and serialization
dataclasses>=0.8; python_version < "3.7"

# Shared dependencies that the agent needs to understand the project structure
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0

# Database connectivity (for understanding the generated services)
asyncpg>=0.29.0
redis[hiredis]>=5.0.0

# HTTP client for testing generated services
httpx>=0.25.0

# Configuration management (Hydra)
hydra-core>=1.3.0
omegaconf>=2.3.0

# Testing framework (for generating tests)
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Code quality tools (for understanding best practices)
black>=23.11.0
isort>=5.12.0
flake8>=6.1.0

# Logging and monitoring
structlog>=23.2.0

# AI integration (for future enhancements)
anthropic>=0.8.0

# Template rendering (if needed for complex functionality)
jinja2>=3.1.0

# Authentication and security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# Email functionality  
pydantic[email]>=2.5.0

# File processing
PyPDF2>=3.0.0
python-docx>=0.8.11
python-pptx>=0.6.21
pillow>=10.0.0
aiofiles>=24.1.0

# Excel/CSV processing
openpyxl>=3.1.0
pandas>=2.1.0

# Additional testing dependencies
pytest-cov>=4.1.0
pytest-mock>=3.12.0
coverage>=7.3.0
faker>=20.1.0

# Selenium and Chrome WebDriver for browser automation testing
selenium>=4.15.0
webdriver-manager>=4.0.1
chromedriver-py>=118.0.5993.70

# PyUnit (unittest) - included in Python standard library but with extensions
unittest-xml-reporting>=3.2.0
nose2>=0.14.0
parameterized>=0.9.0

# Additional browser testing and automation
playwright>=1.40.0
beautifulsoup4>=4.12.0
requests-html>=0.10.0

# Test data generation and utilities
factory-boy>=3.3.0
freezegun>=1.2.2
responses>=0.24.0
vcrpy>=5.1.0

# Performance and load testing enhancements
pytest-benchmark>=4.0.0
pytest-xdist>=3.5.0  # Parallel test execution
pytest-timeout>=2.2.0

# Browser testing utilities
pytest-selenium>=4.1.0
pytest-html>=4.1.0  # HTML test reports
allure-pytest>=2.13.2  # Advanced test reporting

# Performance and monitoring
locust>=2.17.0

# Docker SDK (for lab containers)
docker>=6.1.0

# Additional FastAPI dependencies
python-dotenv>=1.0.0
starlette>=0.27.0

# PostgreSQL testing utilities
pytest-postgresql>=5.0.0

# Redis testing
fakeredis>=2.20.1

# Additional testing and development tools
tox>=4.11.4  # Testing across multiple Python versions
pre-commit>=3.6.0  # Git pre-commit hooks for code quality
bandit>=1.7.5  # Security testing
safety>=3.0.1  # Dependency vulnerability scanning

# Memory and profiling tools for performance testing
memory-profiler>=0.61.0
psutil>=5.9.6

# API testing utilities
httpx>=0.25.0  # Already included above but ensuring it's here
aioresponses>=0.7.6  # Async HTTP mocking
